Replace getRankings.php
Prepared the stats service for keycloak integration, mysql env variables, etc.
Note mysql connectivity has not been tested. This will be tested with the subtasks to this (#66-#69)
Edited widget.rankings.js to call the endpoint instead of the php. This is the only place getRankings.php is referenced.
Moved the four class .php files to the stats subdir. In subtasks #66-#69 these will be replaced by .py class definitions which will then be imported into views.py and instantiated by the /getRankings endpoint to generate the proper response dict/JSON