Implemented List of Shame
Implemented shameList endpoint in cardbox service. This is exposed to the client.
Updated the front end to call the shameList endpoint and use the token in the header
Added a returnValidAlphas endpoint to the lexicon service. This takes in a list of alphas and returns the ones that have valid solutions.
Moved a couple other helper functions from xerafinLib to the cardbox service.